backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
jira==3.5.2
python-gitlab==4.4.0
atlassian-python-api==3.41.0
cachetools==5.3.2
